requirements.txt
anthropic>=0.25.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-docx>=0.8.11
PyPDF2>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0